+++ CHANGES.md
@@ -44,6 +44,11 @@
   color bar management in xcube viewer,
   see https://github.com/xcube-dev/xcube-viewer/issues/390. (#1043)  
 
+### Other changes
+
+* Renamed internal color mapping types from `"node"`, `"bound"`, `"key"` 
+  into `"continuous"`, `"stepwise"`, `"categorical"`.
+
 ## Changes in 1.6.0
 
 ### Enhancements